Update target framework, add null checks, set default values
Updated the target framework in `Berrevoets.Play.Economy.Common.csproj` to `net8.0` and incremented the package version to `1.0.1`. Added a null-forgiving operator (`!`) to the `database` variable in `Extensions.cs` to ensure it is not null. Initialized `DatabaseName` and `CollectionName` properties in `ServiceSettings.cs` with default values of `string.Empty` to prevent null values.